refactor: consolidate campaign tools and fix import paths
Move all campaign-related tools from campaignTools.ts into campaign.ts to
improve code organization. Update all import paths throughout the codebase
to reflect the new file structure. Convert dynamic imports to static
imports where appropriate for better performance and maintainability.

- Move campaignTools.ts content to campaign.ts
- Update import paths from ./campaignTools to ./campaign
- Convert redundant dynamic imports to static imports
- Apply Biome formatting fixes across 20 files
- Maintain full functionality while improving code structure

This consolidation eliminates code duplication and provides a cleaner,
more maintainable codebase structure for campaign management tools.
