+++ src/components/pdf-upload/PdfList.tsx
@@ -1,8 +1,7 @@
 import { useCallback, useEffect, useState } from "react";
-import { authenticatedFetchWithExpiration } from "../../lib/auth";
-import { getStoredJwt } from "../../lib/auth";
+import { ERROR_MESSAGES, USER_MESSAGES } from "../../constants";
+import { authenticatedFetchWithExpiration, getStoredJwt } from "../../lib/auth";
 import { API_CONFIG } from "../../shared";
-import { USER_MESSAGES, ERROR_MESSAGES } from "../../constants";
 
 interface PdfFile {
   fileKey: string;
