+++ src/lib/toolAuth.ts
@@ -1,5 +1,5 @@
-import { createAuthHeaders } from "./auth";
 import { ERROR_MESSAGES } from "../constants";
+import { createAuthHeaders } from "./auth";
 
 /**
  * Helper function for tools to create authenticated fetch requests
