+++ src/tools/index.ts
@@ -3,16 +3,12 @@
  * Tools can either require human confirmation or execute automatically
  */
 
-import { getCurrentAgent } from "agents";
-import { unstable_scheduleSchema } from "agents/schedule";
-import { tool } from "ai";
-import { z } from "zod";
-import type { Chat } from "../server";
-
 // Campaign-related tools
-import { campaignTools as importedCampaignTools } from "./campaignTools";
+import { campaignTools as importedCampaignTools } from "./campaign";
 // PDF-related tools have been moved to ./tools/pdfTools.ts
-import { pdfTools as importedPdfTools } from "./pdfTools";
+import { pdfTools as importedPdfTools } from "./pdf";
+// General tools
+import { tools as generalTools } from "./tools";
 
 console.log("DEBUG importedCampaignTools:", importedCampaignTools);
 console.log("DEBUG createCampaign tool:", importedCampaignTools.createCampaign);
@@ -25,89 +21,12 @@ console.log(
   importedCampaignTools.createCampaign.parameters
 );
 
-const scheduleTask = tool({
-  description: "A tool to schedule a task to be executed at a later time",
-  parameters: unstable_scheduleSchema,
-  execute: async ({ when, description }) => {
-    // we can now read the agent context from the ALS store
-    const { agent } = getCurrentAgent<Chat>();
-
-    function throwError(msg: string): string {
-      throw new Error(msg);
-    }
-    if (when.type === "no-schedule") {
-      return "Not a valid schedule input";
-    }
-    const input =
-      when.type === "scheduled"
-        ? when.date // scheduled
-        : when.type === "delayed"
-          ? when.delayInSeconds // delayed
-          : when.type === "cron"
-            ? when.cron // cron
-            : throwError("not a valid schedule input");
-    try {
-      agent!.schedule(input!, "executeTask", description);
-    } catch (error) {
-      console.error("error scheduling task", error);
-      return `Error scheduling task: ${error}`;
-    }
-    return `Task scheduled for type "${when.type}" : ${input}`;
-  },
-});
-
-/**
- * Tool to list all scheduled tasks
- * This executes automatically without requiring human confirmation
- */
-const getScheduledTasks = tool({
-  description: "List all tasks that have been scheduled",
-  parameters: z.object({}),
-  execute: async () => {
-    const { agent } = getCurrentAgent<Chat>();
-
-    try {
-      const tasks = agent!.getSchedules();
-      if (!tasks || tasks.length === 0) {
-        return "No scheduled tasks found.";
-      }
-      return tasks;
-    } catch (error) {
-      console.error("Error listing scheduled tasks", error);
-      return `Error listing scheduled tasks: ${error}`;
-    }
-  },
-});
-
-/**
- * Tool to cancel a scheduled task by its ID
- * This executes automatically without requiring human confirmation
- */
-const cancelScheduledTask = tool({
-  description: "Cancel a scheduled task using its ID",
-  parameters: z.object({
-    taskId: z.string().describe("The ID of the task to cancel"),
-  }),
-  execute: async ({ taskId }) => {
-    const { agent } = getCurrentAgent<Chat>();
-    try {
-      await agent!.cancelSchedule(taskId);
-      return `Task ${taskId} has been successfully canceled.`;
-    } catch (error) {
-      console.error("Error canceling scheduled task", error);
-      return `Error canceling task ${taskId}: ${error}`;
-    }
-  },
-});
-
 /**
  * Export all available tools
  * These will be provided to the AI model to describe available capabilities
  */
 export const tools = {
-  scheduleTask,
-  getScheduledTasks,
-  cancelScheduledTask,
+  ...generalTools,
   ...importedPdfTools,
   ...importedCampaignTools,
 };
@@ -123,8 +42,7 @@ import type { ToolExecutionOptions } from "ai";
 
 export const executions = {
   createCampaign: async (params: { name: string; jwt?: string }) => {
-    const { campaignTools } = await import("./campaignTools");
-    return campaignTools.createCampaign.execute(
+    return importedCampaignTools.createCampaign.execute(
       params,
       {} as ToolExecutionOptions
     );
@@ -133,8 +51,7 @@ export const executions = {
     campaignId: string;
     jwt?: string;
   }) => {
-    const { campaignTools } = await import("./campaignTools");
-    return campaignTools.listCampaignResources.execute(
+    return importedCampaignTools.listCampaignResources.execute(
       params,
       {} as ToolExecutionOptions
     );
@@ -146,15 +63,13 @@ export const executions = {
     resourceName?: string;
     jwt?: string;
   }) => {
-    const { campaignTools } = await import("./campaignTools");
-    return campaignTools.addResourceToCampaign.execute(
+    return importedCampaignTools.addResourceToCampaign.execute(
       params,
       {} as ToolExecutionOptions
     );
   },
   showCampaignDetails: async (params: { campaignId: string; jwt?: string }) => {
-    const { campaignTools } = await import("./campaignTools");
-    return campaignTools.showCampaignDetails.execute(
+    return importedCampaignTools.showCampaignDetails.execute(
       params,
       {} as ToolExecutionOptions
     );
