+++ tests/campaign/campaignTools.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { campaignTools } from "../../src/tools/campaignTools";
 import { AUTH_CODES, USER_MESSAGES } from "../../src/constants";
+import { campaignTools } from "../../src/tools/campaignTools";
 
 // Mock fetch globally
 global.fetch = vi.fn();
