+++ tests/tools/definitions.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { tools } from "../../src/tools";
+import { tools } from "../../src/tools/tools";
 
 /**
  * Tool Definitions Test Suite
