+++ worker-configuration.d.ts
@@ -350,8 +350,7 @@ declare const performance: Performance;
 declare const Cloudflare: Cloudflare;
 declare const origin: string;
 declare const navigator: Navigator;
-interface TestController {
-}
+type TestController = {}
 interface ExecutionContext {
     waitUntil(promise: Promise<any>): void;
     passThroughOnException(): void;
@@ -2022,8 +2021,7 @@ interface TraceItem {
 interface TraceItemAlarmEventInfo {
     readonly scheduledTime: Date;
 }
-interface TraceItemCustomEventInfo {
-}
+type TraceItemCustomEventInfo = {}
 interface TraceItemScheduledEventInfo {
     readonly scheduledTime: number;
     readonly cron: string;
@@ -6276,8 +6274,7 @@ declare abstract class D1PreparedStatement {
 // but this will ensure type checking on older versions still passes.
 // TypeScript's interface merging will ensure our empty interface is effectively
 // ignored when `Disposable` is included in the standard lib.
-interface Disposable {
-}
+type Disposable = {}
 /**
  * An email message that can be sent from a Worker.
  */
@@ -6725,8 +6722,7 @@ declare namespace Rpc {
     };
 }
 declare namespace Cloudflare {
-    interface Env {
-    }
+    type Env = {}
 }
 declare module 'cloudflare:workers' {
     export type RpcStub<T extends Rpc.Stubable> = Rpc.Stub<T>;
